Reducing the number of warnings from clang compiler

+++ src/common/hhfft_common_avx_f.h
@@ -56,6 +56,7 @@ inline ComplexF load_real_F(const float *v)
 inline void store_F(ComplexF val, float &r, float &i)
 {
     float v[2];
+    // Note according to intriniscs guide, memory address does not need to be aligned
     _mm_storel_pi((__m64*) v, val);
     r = val[0]; i = val[1];
 }
@@ -155,7 +156,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF k0 = broadcast32_F(0.5);
-        const ComplexF k1 = broadcast32_F(0.5*sqrt(3.0));
+        const ComplexF k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         ComplexF t0 = x_in[1] + x_in[2];
         ComplexF t1 = x_in[0] - k0*t0;
         ComplexF t2 = mul_i_F(k1*(x_in[1] - x_in[2]));
@@ -195,10 +196,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF k1 = broadcast32_F(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF k2 = broadcast32_F(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF k3 = broadcast32_F(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF k4 = broadcast32_F(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF k1 = broadcast32_F(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF k2 = broadcast32_F(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF k3 = broadcast32_F(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF k4 = broadcast32_F(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF t0 = x_in[1] + x_in[4];
         ComplexF t1 = x_in[2] + x_in[3];
@@ -233,9 +234,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF k0 = broadcast32_F(0.5);
         ComplexF k1;
         if (forward)
-            k1 = broadcast32_F(0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F(-0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(-0.5*sqrt(3.0)));
 
         ComplexF t6 = x_in[2] + x_in[4];
         ComplexF t7 = x_in[1] + x_in[5];
@@ -262,7 +263,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF k = broadcast32_F(sqrt(0.5));
+        const ComplexF k = broadcast32_F(float(sqrt(0.5)));
 
         ComplexF t12 = x_in[1] + x_in[5];
         ComplexF t13 = x_in[3] + x_in[7];
@@ -608,7 +609,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF2 k0 = broadcast32_F2(0.5);
-        const ComplexF2 k1 = broadcast32_F2(0.5*sqrt(3.0));
+        const ComplexF2 k1 = broadcast32_F2(float(0.5*sqrt(3.0)));
         ComplexF2 t0 = x_in[1] + x_in[2];
         ComplexF2 t1 = x_in[0] - k0*t0;
         ComplexF2 t2 = mul_i_F2(k1*(x_in[1] - x_in[2]));
@@ -648,10 +649,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF2 k1 = broadcast32_F2(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF2 k2 = broadcast32_F2(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF2 k3 = broadcast32_F2(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF2 k4 = broadcast32_F2(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF2 k1 = broadcast32_F2(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF2 k2 = broadcast32_F2(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF2 k3 = broadcast32_F2(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF2 k4 = broadcast32_F2(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF2 t0 = x_in[1] + x_in[4];
         ComplexF2 t1 = x_in[2] + x_in[3];
@@ -686,9 +687,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF2 k0 = broadcast32_F2(0.5);
         ComplexF2 k1;
         if (forward)
-            k1 = broadcast32_F2(0.5*sqrt(3.0));
+            k1 = broadcast32_F2(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F2(-0.5*sqrt(3.0));
+            k1 = broadcast32_F2(float(-0.5*sqrt(3.0)));
 
         ComplexF2 t6 = x_in[2] + x_in[4];
         ComplexF2 t7 = x_in[1] + x_in[5];
@@ -715,7 +716,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF2 k = broadcast32_F2(sqrt(0.5));
+        const ComplexF2 k = broadcast32_F2(float(sqrt(0.5)));
 
         ComplexF2 t12 = x_in[1] + x_in[5];
         ComplexF2 t13 = x_in[3] + x_in[7];
@@ -1102,7 +1103,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF4 k0 = broadcast32_F4(0.5);
-        const ComplexF4 k1 = broadcast32_F4(0.5*sqrt(3.0));
+        const ComplexF4 k1 = broadcast32_F4(float(0.5*sqrt(3.0)));
         ComplexF4 t0 = x_in[1] + x_in[2];
         ComplexF4 t1 = x_in[0] - k0*t0;
         ComplexF4 t2 = mul_i_F4(k1*(x_in[1] - x_in[2]));
@@ -1142,10 +1143,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF4 k1 = broadcast32_F4(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF4 k2 = broadcast32_F4(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF4 k3 = broadcast32_F4(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF4 k4 = broadcast32_F4(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF4 k1 = broadcast32_F4(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF4 k2 = broadcast32_F4(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF4 k3 = broadcast32_F4(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF4 k4 = broadcast32_F4(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF4 t0 = x_in[1] + x_in[4];
         ComplexF4 t1 = x_in[2] + x_in[3];
@@ -1180,9 +1181,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF4 k0 = broadcast32_F4(0.5);
         ComplexF4 k1;
         if (forward)
-            k1 = broadcast32_F4(0.5*sqrt(3.0));
+            k1 = broadcast32_F4(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F4(-0.5*sqrt(3.0));
+            k1 = broadcast32_F4(float(-0.5*sqrt(3.0)));
 
         ComplexF4 t6 = x_in[2] + x_in[4];
         ComplexF4 t7 = x_in[1] + x_in[5];
@@ -1209,7 +1210,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF4 k = broadcast32_F4(sqrt(0.5));
+        const ComplexF4 k = broadcast32_F4(float(sqrt(0.5)));
 
         ComplexF4 t12 = x_in[1] + x_in[5];
         ComplexF4 t13 = x_in[3] + x_in[7];
