+++ src/common/hhfft_common_plain_f.h
@@ -174,7 +174,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF k0 = broadcast32_F(0.5);
-        const ComplexF k1 = broadcast32_F(0.5*sqrt(3.0));
+        const ComplexF k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         ComplexF t0 = x_in[1] + x_in[2];
         ComplexF t1 = x_in[0] - k0*t0;
         ComplexF t2 = mul_i_F(k1*(x_in[1] - x_in[2]));
@@ -214,10 +214,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF k1 = broadcast32_F(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF k2 = broadcast32_F(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF k3 = broadcast32_F(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF k4 = broadcast32_F(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF k1 = broadcast32_F(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF k2 = broadcast32_F(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF k3 = broadcast32_F(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF k4 = broadcast32_F(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF t0 = x_in[1] + x_in[4];
         ComplexF t1 = x_in[2] + x_in[3];
@@ -252,9 +252,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF k0 = broadcast32_F(0.5);
         ComplexF k1;
         if (forward)
-            k1 = broadcast32_F(0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F(-0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(-0.5*sqrt(3.0)));
 
         ComplexF t6 = x_in[2] + x_in[4];
         ComplexF t7 = x_in[1] + x_in[5];
@@ -281,7 +281,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF k = broadcast32_F(sqrt(0.5));
+        const ComplexF k = broadcast32_F(float(sqrt(0.5)));
 
         ComplexF t12 = x_in[1] + x_in[5];
         ComplexF t13 = x_in[3] + x_in[7];
