+++ src/common/hhfft_common_sse2_f.h
@@ -56,6 +56,7 @@ inline ComplexF load_real_F(const float *v)
 inline void store_F(ComplexF val, float &r, float &i)
 {
     float v[2];
+    // It is assumed that there is no requirement for alignment, similar to _mm_loadl_pi
     _mm_storel_pi((__m64*) v, val);
     r = val[0]; i = val[1];
 }
@@ -159,7 +160,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF k0 = broadcast32_F(0.5);
-        const ComplexF k1 = broadcast32_F(0.5*sqrt(3.0));
+        const ComplexF k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         ComplexF t0 = x_in[1] + x_in[2];
         ComplexF t1 = x_in[0] - k0*t0;
         ComplexF t2 = mul_i_F(k1*(x_in[1] - x_in[2]));
@@ -199,10 +200,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF k1 = broadcast32_F(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF k2 = broadcast32_F(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF k3 = broadcast32_F(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF k4 = broadcast32_F(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF k1 = broadcast32_F(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF k2 = broadcast32_F(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF k3 = broadcast32_F(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF k4 = broadcast32_F(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF t0 = x_in[1] + x_in[4];
         ComplexF t1 = x_in[2] + x_in[3];
@@ -237,9 +238,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF k0 = broadcast32_F(0.5);
         ComplexF k1;
         if (forward)
-            k1 = broadcast32_F(0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F(-0.5*sqrt(3.0));
+            k1 = broadcast32_F(float(-0.5*sqrt(3.0)));
 
         ComplexF t6 = x_in[2] + x_in[4];
         ComplexF t7 = x_in[1] + x_in[5];
@@ -266,7 +267,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF k = broadcast32_F(sqrt(0.5));
+        const ComplexF k = broadcast32_F(float(sqrt(0.5)));
 
         ComplexF t12 = x_in[1] + x_in[5];
         ComplexF t13 = x_in[3] + x_in[7];
@@ -613,7 +614,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     if (radix == 3)
     {
         const ComplexF2 k0 = broadcast32_F2(0.5);
-        const ComplexF2 k1 = broadcast32_F2(0.5*sqrt(3.0));
+        const ComplexF2 k1 = broadcast32_F2(float(0.5*sqrt(3.0)));
         ComplexF2 t0 = x_in[1] + x_in[2];
         ComplexF2 t1 = x_in[0] - k0*t0;
         ComplexF2 t2 = mul_i_F2(k1*(x_in[1] - x_in[2]));
@@ -653,10 +654,10 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 5
     if (radix == 5)
     {
-        const ComplexF2 k1 = broadcast32_F2(cos(2.0*M_PI*1.0/5.0));
-        const ComplexF2 k2 = broadcast32_F2(sin(2.0*M_PI*1.0/5.0));
-        const ComplexF2 k3 = broadcast32_F2(-cos(2.0*M_PI*2.0/5.0));
-        const ComplexF2 k4 = broadcast32_F2(sin(2.0*M_PI*2.0/5.0));
+        const ComplexF2 k1 = broadcast32_F2(float(cos(2.0*M_PI*1.0/5.0)));
+        const ComplexF2 k2 = broadcast32_F2(float(sin(2.0*M_PI*1.0/5.0)));
+        const ComplexF2 k3 = broadcast32_F2(float(-cos(2.0*M_PI*2.0/5.0)));
+        const ComplexF2 k4 = broadcast32_F2(float(sin(2.0*M_PI*2.0/5.0)));
 
         ComplexF2 t0 = x_in[1] + x_in[4];
         ComplexF2 t1 = x_in[2] + x_in[3];
@@ -691,9 +692,9 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
         const ComplexF2 k0 = broadcast32_F2(0.5);
         ComplexF2 k1;
         if (forward)
-            k1 = broadcast32_F2(0.5*sqrt(3.0));
+            k1 = broadcast32_F2(float(0.5*sqrt(3.0)));
         else
-            k1 = broadcast32_F2(-0.5*sqrt(3.0));
+            k1 = broadcast32_F2(float(-0.5*sqrt(3.0)));
 
         ComplexF2 t6 = x_in[2] + x_in[4];
         ComplexF2 t7 = x_in[1] + x_in[5];
@@ -720,7 +721,7 @@ template<size_t radix, bool forward> inline __attribute__((always_inline)) void 
     // Implementation for radix = 8
     if (radix == 8)
     {
-        const ComplexF2 k = broadcast32_F2(sqrt(0.5));
+        const ComplexF2 k = broadcast32_F2(float(sqrt(0.5)));
 
         ComplexF2 t12 = x_in[1] + x_in[5];
         ComplexF2 t13 = x_in[3] + x_in[7];
